Replace Markdown file translation script
The commit replaces the Python script `translate-md.py`, which was responsible for translating Markdown files
into multiple languages, with the original translation script, using the original from the tagit repository.
This means that it is always up to date. In addition, some reorganization has taken place,
but this has nothing to do with the functionality of migit.